requirements.txt
Flask
python-dotenv
requests
PyJWT
Flask-Cors